requirements.txt
aiogram==3.0
removebg
python-dotenv